requirements.txt
fastapi
uvicorn
python-multipart
pillow
numpy
opencv-python-headless
ultralytics